requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-dotenv
openai
pandas
numpy
scikit-learn
xgboost
yfinance
